15 exercise: refactor code style

+++ 15_constructorFunctions/assets/js/main.js
@@ -16,31 +16,19 @@ function Calculator() {
     document.addEventListener('click', (e) => {
       const element = e.target;
 
-      if(element.classList.contains('btn-num'))
-        this.btnToDisplay(element.innerText);
-
-      if(element.classList.contains('btn-clear'))
-        this.clearDisplay();
-
-      if(element.classList.contains('btn-del'))
-        this.deleteOne();
-
-      if(element.classList.contains('btn-eq'))
-        this.equal();
+      if(element.classList.contains('btn-num')) this.btnToDisplay(element.innerText);
+      if(element.classList.contains('btn-clear')) this.clearDisplay();
+      if(element.classList.contains('btn-del')) this.deleteOne();
+      if(element.classList.contains('btn-eq')) this.equal();
     });
   };
 
-  this.btnToDisplay = (value) => {
+  this.btnToDisplay = (value) => { 
     this.display.value += value;
+    this.display.focus();
   };
-
-  this.clearDisplay = () => {
-    this.display.value = '';
-  };
-
-  this.deleteOne = () => {
-    this.display.value = this.display.value.slice(0, -1);
-  };
+  this.clearDisplay = () => this.display.value = '';
+  this.deleteOne = () => this.display.value = this.display.value.slice(0, -1);
 
   this.equal = () => {
     let account = this.display.value;
@@ -59,7 +47,6 @@ function Calculator() {
       return;
     }
   };
-
 
 }
